requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
beanie==1.29.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
debugpy==1.8.12
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
environs==14.1.1
fastapi==0.115.7
fastapi-cli==0.0.7
graphql-core==3.2.6
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
libcst==1.6.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
strawberry-graphql==0.258.0
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
